ecommerce/CV/requirements.txt
#Python libraries
pytesseract
Pillow